optimize remove product image

+++ app/src/main/java/aromko/de/wishlist/activity/EditWishActivity.java
@@ -198,7 +198,7 @@ public class EditWishActivity extends AppCompatActivity {
 
         if (ivProductImage.getTag().toString().equals(getString(R.string.txtImageChanged))) {
             isImageSet = true;
-        } else if (ivProductImage.getTag().toString().equals(getString(R.string.txtImageDeleted))) {
+        } else if (ivProductImage.getTag().toString().equals(getString(R.string.txtImageDeleted)) && !"".equals(photoUrl)) {
             photoHelper.deleteImageFromFirebaseStorage(btnDeleteImage.getTag().toString());
             isImageSet = false;
             photoUrl = "";
@@ -228,7 +228,6 @@ public class EditWishActivity extends AppCompatActivity {
                 .build(this);
         startActivityForResult(intent, AUTOCOMPLETE_REQUEST_CODE);
     }
-
 
 
     private String setLocationText(Place place) {
